+++ handwriting_ocr/data/data_create_sets.py
@@ -1,104 +1,101 @@
+# Copyright 2020 Břetislav Hájek <info@bretahajek.com>
+# Licensed under the MIT License. See LICENSE for details.
+"""Modelu for creating sets (train/dev/test) of normalized images."""
+
 import argparse
-import glob
-import os
-import random
-import sys
-from shutil import copyfile
+from pathlib import Path
 
-import cv2
+import cv2 as cv
 import numpy as np
+from tqdm import tqdm
 
-location = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(os.path.join(location, "../"))
-
-from create_csv import create_csv
-from data_extractor import datasets
-from ocr.viz import print_progress_bar
+from handwriting_ocr.data.data_loader import DATASETS, DATA_FOLDER
+from handwriting_ocr.ocr.normalization import word_normalization
 
 
-random.seed(17)  # Make the datasets split random, but reproducible
-data_folder = "words_final"
-output_folder = os.path.join(location, "../../data/sets/")
+def create_sets(data_path, test_size, dev_size, seed=42):
+    """Loads all data and process them into train/dev/test sets.
 
-# Sets percent distribution
-test_set = 0.1
-validation_set = 0.1
+    It loads all available datasets from data_path, splits them into train/dev/test sets
+    and normalize images. Labels are saved into labels.txt file as
+    `{filename.png}\t{label}` (separated by tab)
+
+    Args:
+        data_path (Path): Data folder path
+        test_size (float): Percentage of test images from all images (between 0-1)
+        dev_size (float): Percentage of dev images from all images (between 0-1)
+        seed (int): Seed value for reproducibility of split
+    """
+    np.random.seed(seed)
+    lines = []
+    for d in DATASETS:
+        if d.is_downloaded(data_path):
+            lines.extend(d.load(data_path))
+    np.random.shuffle(lines)
+
+    test_i, dev_i = (len(lines) * np.array([test_size, test_size + dev_size])).astype(
+        int
+    )
+
+    sets = {
+        "test": lines[:test_i],
+        "dev": lines[test_i:dev_i],
+        "train": lines[dev_i:],
+    }
+
+    for k, set_lines in sets.items():
+        print(f"{k} images: {len(set_lines)}")
+
+        folder = data_path / "sets" / k
+        folder.mkdir(parents=True, exist_ok=True)
+
+        label_file = folder.joinpath("labels.txt").open("w+")
+        for i, (path, label) in enumerate(tqdm(set_lines)):
+            image = cv.imread(str(path))
+            if image.shape[0] < 15:
+                continue
+
+            norm = word_normalization(
+                image, height=64, border=False, tilt=False, hyst_norm=False
+            )
+
+            out_name = f"image_{i:05}.png"
+            cv.imwrite(str(folder.joinpath(out_name)), norm)
+            label_file.write(f"{out_name}\t{label}\n")
+
+        label_file.close()
 
 
-parser = argparse.ArgumentParser(
-    description="Script sliting processed words into train, validation and test sets."
-)
-parser.add_argument(
-    "-d",
-    "--dataset",
-    nargs="*",
-    choices=datasets.keys(),
-    help="Pick dataset(s) to be used.",
-)
-parser.add_argument(
-    "-p",
-    "--path",
-    nargs="*",
-    default=[],
-    help="""Path to folder containing the dataset. For multiple datasets
-    provide path or ''. If not set, default paths will be used.""",
-)
-parser.add_argument(
-    "--output",
-    default="data-handwriting/sets",
-    help="Directory for normalized and split data",
-)
-parser.add_argument(
-    "--csv",
-    action="store_true",
-    default=False,
-    help="Include flag if you want to create csv files along with split.",
-)
+def get_args():
+    """ArgumentParser for sets creation."""
+    parser = argparse.ArgumentParser(
+        description="Script for creating sets (train/dev/test) of normalized images."
+    )
+    parser.add_argument(
+        "--test_size",
+        type=float,
+        default=0.1,
+        help="Percentage (0-1) size of test set.",
+    )
+    parser.add_argument(
+        "--dev_size",
+        type=float,
+        default=0.1,
+        help="Percentage (0-1) size of development set.",
+    )
+    parser.add_argument("--seed", type=int, default=42, help="Seed for random shuffle.")
+    parser.add_argument(
+        "--data_path",
+        type=Path,
+        default=DATA_FOLDER,
+        help="Path of data folder (default is recommended).",
+    )
+    return parser.parse_args()
 
 
 if __name__ == "__main__":
-    args = parser.parse_args()
-    if args.dataset == ["all"]:
-        args.dataset = list(datasets.keys())[:-1]
+    args = get_args()
+    for d in DATASETS:
+        d.download(DATA_FOLDER)
 
-    assert args.path == [] or len(args.dataset) == len(
-        args.path
-    ), "provide same number of paths as datasets (use '' for default)"
-    if args.path != []:
-        for ds, path in zip(args.dataset, args.path):
-            datasets[ds][1] = path
-
-    if not os.path.exists(output_folder):
-        os.makedirs(output_folder)
-
-    imgs = []
-    for ds in args.dataset:
-        for loc, _, _ in os.walk(datasets[ds][1].replace("raw", "processed")):
-            imgs += glob.glob(os.path.join(loc, "*.png"))
-
-    imgs.sort()
-    random.shuffle(imgs)
-
-    length = len(imgs)
-    sp1 = int((1 - test_set - validation_set) * length)
-    sp2 = int((1 - test_set) * length)
-    img_paths = {"train": imgs[:sp1], "dev": imgs[sp1:sp2], "test": imgs[sp2:]}
-
-    i = 0
-    for split in ["train", "dev", "test"]:
-        split_output = os.path.join(output_folder, split)
-        if not os.path.exists(split_output):
-            os.mkdir(split_output)
-        for im_path in img_paths[split]:
-            copyfile(im_path, os.path.join(split_output, os.path.basename(im_path)))
-            if "_gaplines" in im_path:
-                im_path = im_path[:-3] + "txt"
-                copyfile(im_path, os.path.join(split_output, os.path.basename(im_path)))
-
-            print_progress_bar(i, length)
-            i += 1
-
-        print("\n\tNumber of %s words: %s" % (split, len(os.listdir(split_output))))
-
-    if args.csv:
-        create_csv(output_folder)
+    create_sets(args.data_path, args.test_size, args.dev_size, args.seed)
